Add files via upload
In this script I try to perform the krigging interpolation using elevation as an additional input for the safe chilling values of South America. The methods were mostly taken and Modified from: Luedeling & Benmoussa (2020): Supplementary materials for the paper “Severe winter chill decline impacts Tunisian fruit and nut orchards” - Additional information and analysis code. The biggest problem is, that sofar it appears that only one point is used for the krigging method