Update create_maps.R
Hi everyone,
I cleared one of the problems when doing the interpolation. When creating the spatial point data frame 'P' the wrong projection format was used. In the changed version WGS84 is used and in the second step the object is then converted to the desired planar projection mode.

The inverse distance weighing works fine now. However, the krigging still yields weird values. The outcome is totally out of the desired range. @EduardoFernandezC please have a look at that part (code line 99 - 142). The fit of the variogram is not great but at least reasonable, but somehow the values are still completely wrong. It would be great if you can have a look at it.

Thanks a lot and see you next week (^_^)/